`Refactor login page and related components`
