+++ src/modules/log_in/components/LoginCard.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import '../styles/LoginCard.css';
+
+interface LoginCardProps {
+  children: React.ReactNode;
+}
+
+export default function LoginCard({ children }: LoginCardProps) {
+  return (
+    <div className="login-card">
+      <div className="blob" />
+      {children}
+    </div>
+  );
+}
