+++ src/modules/log_in/components/LoginHeader.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import '../styles/LoginHeader.css';
+
+export default function LoginHeader() {
+  return (
+    <header className="login-header">
+      <h1 className="login-title">Bienvenid@ a KOLLA</h1>
+      <p className="login-subtitle">Inicia sesión para continuar</p>
+    </header>
+  );
+}
