+++ src/modules/log_in/components/LoginShell.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import '../styles/LoginShell.css';
+
+interface LoginShellProps {
+  children: React.ReactNode;
+}
+
+export default function LoginShell({ children }: LoginShellProps) {
+  return (
+    <div className="login-shell">
+      {children}
+    </div>
+  );
+}
