+++ src/modules/log_in/styles/LoginCard.css
@@ -1,0 +1,48 @@
+.login-card {
+  position: relative;
+  width: min(440px, 92vw);
+  background: var(--surface);
+  border-radius: var(--radius);
+  box-shadow: var(--shadow-lg);
+  overflow: hidden;
+  padding: 28px;
+  border: 1px solid rgba(27, 97, 42, 0.08);
+  transform: translateZ(0);
+}
+
+/* Borde animado sutil */
+.login-card::before {
+  content: "";
+  position: absolute;
+  inset: -2px;
+  border-radius: calc(var(--radius) + 2px);
+  background: conic-gradient(
+    from 180deg,
+    rgba(43,191,89,0.0),
+    rgba(43,191,89,0.35),
+    rgba(43,191,89,0.0) 65%
+  );
+  filter: blur(10px);
+  z-index: -1;
+  animation: rotateGlow 7s linear infinite;
+  pointer-events: none;
+}
+@keyframes rotateGlow {
+  to { transform: rotate(360deg); }
+}
+
+/* Blob decorativo */
+.blob {
+  position: absolute;
+  width: 380px; height: 380px;
+  border-radius: 50%;
+  background: radial-gradient(circle at 30% 30%, rgba(43,191,89,0.25), rgba(27,97,42,0.05) 60%, transparent 70%);
+  top: -140px; right: -120px;
+  filter: blur(6px);
+  animation: floaty 12s ease-in-out infinite;
+  z-index: 0;
+}
+@keyframes floaty {
+  0%,100% { transform: translateY(0) translateX(0); }
+  50% { transform: translateY(12px) translateX(-8px); }
+}
